Auto stash before checking out "origin/V2"
