+++ .env
@@ -2,10 +2,13 @@ REACT_APP_APPNAME='Axial'
 REACT_APP_NODE_ADDRESS='https://api.avax.network'
 #'http://localhost:8545'
 REACT_APP_NETWORK=mainnet
+REACT_APP_CHAIN_ID=43114
+REACT_APP_HARDHAT_CHAIN_ID=43114
 REACT_APP_ENVIRONMENT='PROD'
 REACT_APP_IPFS_API_URL='https://ipfs.snowapi.net/ipfs/'
 REACT_APP_APIADDR='https://api.snowapi.net/graphql'
 REACT_APP_DEVAPIADDR='https://api.snowapi.net/graphql'
+REACT_APP_BASE_URL='https://base.snowball.io'
 # swap
 REACT_APP_SWAP_ROUTER_ADDRESS='0x05F2F176b491ec11555cbE45B99748F999E13933'
 # analytics
@@ -32,3 +35,4 @@ REACT_APP_FEE_DISTRIBUTOR_CONTRACT_ADDRESS='0xAd86ef5fD2eBc25bb9Db41A1FE8d0f2a32
 #gaugeProxy
 REACT_APP_GAUGE_PROXY_ADDRESS='0x215D5eDEb6A6a3f84AE9d72962FEaCCdF815BF27'
 #vote-allocation
+REACT_APP_AXIAL_MASTERCHEF_CONTRACT_ADDRESS='0x958C0d0baA8F220846d3966742D4Fb5edc5493D3'
