+++ src/app/containers/BlockChain/Web3/index.tsx
@@ -17,13 +17,11 @@ import { BaseProvider, getDefaultProvider } from "@ethersproject/providers";
 import { Web3Provider } from "@ethersproject/providers";
 import { NetworkContextName } from "../constants";
 import { rpcUrl } from "../utils/wallet/connectors";
-import { ChainId } from "app/constants";
 import { Web3ReactContextInterface } from "@web3-react/core/dist/types";
 import Web3ReactManager from "./Web3ReactManager";
 const Web3ProviderNetwork = createWeb3ReactRoot(NetworkContextName);
 
 export function useActiveWeb3React(): Web3ReactContextInterface<Web3Provider> & {
-  chainId?: ChainId;
 } {
   const context = useWeb3React<Web3Provider>();
   const contextNetwork = useWeb3React<Web3Provider>(NetworkContextName);
