+++ src/config/index.ts
@@ -2,7 +2,6 @@ import { env } from "environment";
 
 const IS_MAINNET = env.NETWORK === "mainnet";
 
-const C_CHAIN_ID = IS_MAINNET ? 43114 : 43113;
 
 const MAX_RETRIES = 5;
 
@@ -50,4 +49,4 @@ const WAVAX = "0xB31f66AA3C1e785363F0875A1B74E27b85FD66c7";
 const PNG = "0x60781C2586D68229fde47564546784ab3fACA982";
 const JOE = "0x6e84a6216ea6dacc71ee8e6b0a5b7322eebc0fdd";
 
-export { IS_MAINNET, C_CHAIN_ID, CONTRACTS, MAX_RETRIES, WAVAX, PNG, JOE };
+export { IS_MAINNET, CONTRACTS, MAX_RETRIES, WAVAX, PNG, JOE };
