+++ src/hooks/onlineStatusHook/provider.tsx
@@ -1,4 +1,3 @@
-import { BaseUrl } from "app/constants";
 import React, {
   useState,
   useEffect,
@@ -7,7 +6,9 @@ import React, {
   createContext,
 } from "react";
 
-const PING_RESOURCE = `${BaseUrl}/check`;
+
+
+const PING_RESOURCE = `${process.env.REACT_APP_BASE_URL}/check`;
 const TIMEOUT_TIME_MS = 4000;
 const onlinePollingInterval = 60 * 1000;
